Retransmission buffer may be empty on ACK frame processing
Summary:
There is a race condition with min data frame (skip) and ack frame when it comes to retransmission buffer.
Sometimes, we receive both min data and ack frames, the min data frame gets processed first and it clears out retransmission buffer (expected), but then we get to process the ack frame and there is that CHECK that expects the retransmission buffer to have at least one unacked buffer in it - there is none and then the app crashes.

Reviewed By: siyengar

Differential Revision: D14908630

fbshipit-source-id: d8bcd70291e1fd3618065fdeb44bedc6ac66bcaa