ExternalProject: log to stdout instead of files (#8)
Summary:
If you specify the `LOG_*` directives, then nothing appears on stdout
while the external project is building.  On my Mac, this means it sits
at the Fizz build step for 4+ minutes with no visible progress in the
terminal.

Removing the `LOG_*` directives means that cmake will display all the
output to stdout instead of to files.  This is consistent with the
build behavior for Folly (which is not a cmake ExternalProject).  If
logging the external project builds is desired, the user can capture
the output of the build_helper.sh.

Signed-off-by: Jeff Squyres <jeff@squyres.com>
Pull Request resolved: https://github.com/facebookincubator/mvfst/pull/8

Reviewed By: lnicco

Differential Revision: D15267875

Pulled By: udippant

fbshipit-source-id: 138c0f1af6679d2c6e5bbb243d7891211763820c