fbcode_builder: getdeps: fboss: add OpenNSL fetcher
Summary:
This is towards getting open source FBOSS to build using fbcode_builder.
OpenNSL is one of the dependencies for FBOSS.

OpenNSL is github repository with headers and prebuilt opaque library. Thus, we
can't/don't need to build it. Thus, OpenNSL manifest uses NopBuilder.

fbcode_builder expects the library to be in installed/lib and headers to be in
installed/include, but OpenNSL lib structure is different.  Thus, add explicit
[install.fields] rule for bin/wedge => lib. Once a rule(s) is provided, we must
provide rule(s) for all, so add rule for include => include as well.

Reviewed By: wez

Differential Revision: D15593639

fbshipit-source-id: facb36ff7b9a31f5952ecbc1a9fdb834c0e68d11