Delete default constructor for ShortHeader
Summary:
Running into build failures for the new xcode11 sdk for multiple apps (we're not yet building these on CI, so this is not urgent).

ConnectionId does not have an accessible default constructor, so this means ShortHeader can not be default constructed.

This ends up giving us this build error under the clang++ that comes with the new xcode11 sdk. It's not clear exactly
why this wasn't an error in the clang++ that ships with xcode10.2.1. Swapping out just the clang++ binary from the sdk does resolve the issue (as in, i've isolated the issue particularly to clang++).

```
buck-out/../xplat/quic/codec/Types.h:721:3: error: explicitly defaulted default constructor is implicitly deleted [-Werror,-Wdefaulted-function-deleted]
  ShortHeader() = default;
  ^
buck-out/../xplat/quic/codec/Types.h:731:16: note: default constructor of 'ShortHeader' is implicitly deleted because field 'connectionId_' has an inaccessible default constructor
  ConnectionId connectionId_;
               ^
1 error generated.
```

The fix here is to delete the default constructor of ShortHeader as it is not being used. Deleting this default constructor makes the build succeed.

Reviewed By: JunqiWang

Differential Revision: D16195639

fbshipit-source-id: b6fb14032da5329c61ce327200c82a60f3e1f87a