fbzmq: disable build on darwin and windows
Summary:
boost.fiber is not available on macos and is required by fbzmq,
so disable the build on darwin.

In D17128935 I tried to make libzmq build with cmake to build
on windows, but there's a problem with its cmake and it fails
to configure.

So let's just nop out the build for !linux for now.

Reviewed By: jstrizich

Differential Revision: D17129362

fbshipit-source-id: db89010d18707a4105b8d395b1758f738336495d