Remove canBePaced API from CongestionController
Summary: Centralize this logic into the pacer interface. But instead of having an explicit canBePaced() function (and we already have a dozen other canBePaced functions or boolean flags in other layers), the pacer just use 0us and default write batch size per loop to answer pacing rate queries when it doesn't meet the condition to properly pace the transport writes.

Reviewed By: mjoras

Differential Revision: D16912631

fbshipit-source-id: 0343b126ca5de315af6823067316526b1004cc6c