Reverse linear search for outstanding packet insert point
Summary: Beside the beginning of a connection where we do handshake and have packets from all 3 spaces can possibly in the outstanding packet list, for most parf of the connection, this list only has packets from AppData space. That means the packet numbers are strictly ascending for most parf of a connection. Then linear search from back can potentially reduce the number of times we do the packet number compare.

Reviewed By: mjoras

Differential Revision: D17500221

fbshipit-source-id: 15eafa439e889e45d3e4a84d9e9dad0e2c743c9d