Use array instead of vector for connid
Summary:
In perf traces we see that the allocations of the vector memory shows up.

The maximum connid size is 20 bytes. We previously used a vector because
we thought that it might save some space, however the default size of a vector
is actually 24 bytes.

This diff replaces the vector with an array of size Maximum connid size. This
should keep most code the same, and now will only consume 20 + 1 byte + some alignment bytes.
This should make all the vector memory allocations go away during parsing.

Reviewed By: yangchi

Differential Revision: D17806689

fbshipit-source-id: 041d5f18624fb34ce53534c75b2bdca3513fa107