Add replaceSocket() method to QuicClientTransport
Summary:
Support migration by default on HQServer.
Allow configurable setting to initiate client migration. If client migration is
enabled, we switch sockets on the second request.

We cannot switch sockets right away on connectSuccess, because even though the
1RTT clientWriteCipher is available, the handshake is still not fully complete.
The server would then receive a packet from a different peer address, but it
wouldn't be in the app data packet space, causing check in ServerStateMachine
to fail.

Reviewed By: JunqiWang

Differential Revision: D18246022

fbshipit-source-id: 21cfa9080a703ae09574711206b8f026282b6a34