Remove loss reduction hack in both Cubic and BBR
Summary:
Cubic has a hack that sets a "connection emulation". With this, on
packet loss, it reduces cwnd by 10% instead of 20%. Then a even more aggresive
hack was added into BBR to make the cwnd on par with Cubic when there is packet
loss. This was used to facilitate BBR and Cubic tests when they were tested
side-by-side in the same network. After that, neither of them were sound ideas.

Reviewed By: mjoras

Differential Revision: D17053330

fbshipit-source-id: 17f74f58d7a036f85aaced10ff6d9646abae9e85