Remove open stream sorting.
Summary:
The idea of having a sorted deque to do this makes some sense, but it ends up extremely inefficient for workloads that churn through a lot of streams.

Mitigate this by instead tracking the open streams with a `FastSet`. For a very small number of streams this will probably be slightly less efficient, but not measurably.

Reviewed By: kvtsoy

Differential Revision: D18409885

fbshipit-source-id: 2214640682926b4b4525700df85627f6125cabe8