Fix Quic BBR ProbeRtt
Summary:
Currently when a rtt sample expires, a new sample will replace the
value and refresh the timestamp. And then we always use the refreshed timestamp
to check if rtt sample has expired. But because it just got refreshed, the
answer is always false, then we never probe rtt.

This diff saves the rtt expiration result into a variable when we replace the
rtt sample, then answers the minRttExpireed API by returning this value.

Reviewed By: mjoras

Differential Revision: D18727892

fbshipit-source-id: 3907f705841f37f55c46950ed4e6c24cef50de8f