Replace IOBufQueue in writeBuffer
Summary: This isn't a huge source of cycles, but the flushing of the cache does show up in tperf. Turns out this is very easy to do, and exposed another minor problem which is that we were misusing `clear()` on the `IOBufQueue`. Instead we should have been using `move()` to reset it to being empty.

Reviewed By: siyengar

Differential Revision: D18758922

fbshipit-source-id: 6dcbc29e8bebd93b5787939176799e3df90ed94d