Always refresh AppLimited state in BbrBandwidthSampler when BBR is in ProbeRtt
Summary:
Inside the bandwidth sampler, once AppLimited is set, it lasts until a
packet is sent later than the AppLimited setting time. Currently we only set
AppLimited once when we enter ProbeRtt. This leads to bandwidth sampler exiting
AppLimited state when BBR is still in ProbeRtt. This diff changes to so that
any Ack during ProbeRtt will refresh AppLimited in BandwidthSampler.

Note that the an Ack may exit ProbeRtt. Then BandwidthSampler will be in
AppLimited while BBR will exit ProbeRtt. I think this is the correct behavior of
AppLimited in Bbr BandwidthSampler: Any time it goes to AppLimited, it needs
to mark all packets sent during the next RTT as expericing app-limited
condition. (The so called "bubble" in
https://tools.ietf.org/html/draft-cheng-iccrg-delivery-rate-estimation-00)

Reviewed By: mjoras

Differential Revision: D18886880

fbshipit-source-id: c46279ddc60054dc4cba168c6b054bf2053c69df