collapse most sequential padding frames
Summary:
For padding frames we don't really need to know the exact number of padding frames in the packet,
but just that a padding frame exists.

This diff tries to collapse multiple sequential padding frames into 1 frame while decoding the packet.
While this wont collapse all padding frames, the common case is that padding is sequential and at the
end of the packet.

Reviewed By: JunqiWang

Differential Revision: D18901925

fbshipit-source-id: 17b4eeabdbf4df01c900a9d5078891eb64bf254a