Only wait for one new rtt round during ProbeRtt
Summary:
Currently we need the newRoundTrip boolean to be true twice to exit
ProbeRtt:  once to  set the  probeRttRound_,  and another time to check the
200ms ProbeRtt duration. This makes ProbeRtt really long: it takes more then
one RTT to make this boolean turn True twice; And when it's not true, we don't
check the 200ms ProbeRtt duration, which makes the duration longer than 200ms.
And all of these happen *after* inflight is drained to be a very low value.

This diff changes so that we only wait for one newRoundTrip = True during
ProbeRtt. Once we get to that point, any ack will check the 200ms ProbeRtt
duration.

Reviewed By: bschlinker

Differential Revision: D18899570

fbshipit-source-id: 9a3ff37ea4d237c15859b3dc4db00a761dc29f70