Merge on insert to loss buffer
Summary: Previously we were treating the loss buffer the same way we were the retransmission buffer. That is, each entry represented a single stream frame that was lost. There's actually no reason to do this as we treat it more like the write buffer when retransmitting. It can also lead to pathological cases where we write consecutive byte ranges as separate stream frames.

Reviewed By: yangchi

Differential Revision: D19240352

fbshipit-source-id: 94ffb397fa287688d986a4a91d531d85050d9f98