Fix packet builder header bytes counting
Summary:
In draft-17, we made a change to defer the the packet leng and packet
number writes in long header case in the packet builder. So getHeaderBytes()
API doesn't count these future header bytes. But packet scheduler use this
value to decide how many frames to write into the packet. We end up overwrite a
packet beyond it's size limit.

Reviewed By: mjoras

Differential Revision: D19623496

fbshipit-source-id: 0600cb72f2effb8b4968157e8576cd856d856558