Remove duplicate Quic logging of PARSE_ERROR in 0-rtt buffered case
Summary:
When parsePacket returns anything other than RegularPacket, we already
log the Drop even inside the switch-case blocks before checking if
RegularPacket has been parsed. So the logging of Drop with PARSE_ERROR when the
CodecResult isn't RegularPacket is wrong. For example, right now when we
buffer 0-rtt data, we always log a Drop immediatly afterwards into QLog which
is incorrect.

Reviewed By: JunqiWang

Differential Revision: D20421934

fbshipit-source-id: d836700fd691645951d5e5b49b19cbcc1e5df51a