Avoid more direct usage of [[nodiscard]] in C++14 mode
Summary:
In D20494128 we fixed the direct use of `[[nodiscard]]` in quic, but more cases have apparently gotten into the codebase since. The reason we don't see more errors in fbsource is that clang is lenient allowing this in `-std=c++14` mode, but it causes a mismatch when the flag is honored for MSVC.

Note: quic and other parts of xplat need to stay in `-std=c++14`  mode for the time being. See abandoned diff D20377630 for details on the last attempt to switch to `-std=c++17`.

Reviewed By: yangchi

Differential Revision: D21280601

fbshipit-source-id: 78a1956d644f5f95d892b1fd82c5e899823c8e84